+++ Razor/Razor.csproj
@@ -162,6 +162,7 @@
     <Compile Include="Core\BuffsTimer.cs" />
     <Compile Include="Core\BuffsDebuffs.cs" />
     <Compile Include="Core\BodCapture.cs" />
+    <Compile Include="Core\Dress.cs" />
     <Compile Include="Core\EncodedSpeech.cs" />
     <Compile Include="Core\DamageTracker.cs" />
     <Compile Include="Core\Journal.cs" />
